feature: redirect maya save to silex save
